feature/scores
Added score system - amount of steps is used for score
the fewer steps the better. Uses 'poeng' attribute of Spiller
to keep track of steps used on given maze.

Modified 'move' method in all the LabyrintRuter to update the
steps (Spiller.setPoeng()) in addition to the position.

Scores are serialized to files in the levels directory, with
one seperate file per level. These files are loaded and written
by the Labyrint class.

Added 'scores' menubar item, and window for viewing scores.