patch/windowListener not scaling
Fixed bug were window listener only worked on the first level.
This was because it was set up in Main with a reference to view
but when view was changing (after reloading or going to another level)
the listener was still calling the old reference.

Fixed it by setting up this listener in View (LabViewGrid) rather than
in Main.